New sih/requirements.txt
requests
numpy
pandas
google-generativeai
Flask
aiResponse